backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.0",
    "openai>=2.6.1",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.30.0",
]
